Rename `utils` to `ReactionVariationsUtils`
