+++ app/packs/src/apps/mydb/elements/details/reactions/variationsTab/ReactionVariationsUtils.js
